+++ examples/vue/nuxt-3-catchall/nuxt.config.ts
@@ -1,2 +1,4 @@
 // https://nuxt.com/docs/api/configuration/nuxt-config
-export default defineNuxtConfig({});
+export default defineNuxtConfig({
+  modules: ['@builder.io/sdk-vue/nuxt'],
+});