+++ packages/sdks-tests/src/specs/index.ts
@@ -9,11 +9,15 @@ import {
 } from './custom-breakpoints.js';
 import { CONTENT as dataBindingStyles } from './data-binding-styles.js';
 import { CONTENT as dataBindings } from './data-bindings.js';
+import { DUPLICATE_ATTRIBUTES } from './duplicate-attributes.js';
 import { EDITING_STYLES } from './editing-styles.js';
 import { CONTENT as elementEvents } from './element-events.js';
+import { EXTERNAL_DATA } from './external-data.js';
 import { CONTENT as homepage } from './homepage.js';
 import { CONTENT as image } from './image.js';
 import { INPUT_DEFAULT_VALUE } from './input-default-value.js';
+import { JS_CODE_CONTENT } from './js-code.js';
+import { JS_CONTENT_IS_BROWSER } from './js-content-is-browser.js';
 import { CONTENT as linkUrl } from './link-url.js';
 import { CONTENT as nestedSymbols } from './nested-symbols.js';
 import { CONTENT as reactiveState } from './reactive-state.js';
@@ -36,7 +40,7 @@ function isBrowser(): boolean {
 
 const getPathnameFromWindow = (): string => (isBrowser() ? window.location.pathname : '');
 
-const pages = {
+const PAGES = {
   '/': homepage,
   '/api-version-v1': CONTENT_WITHOUT_SYMBOLS,
   '/api-version-v2': CONTENT_WITHOUT_SYMBOLS,
@@ -46,6 +50,7 @@ const pages = {
   '/css-nesting': cssNesting,
   '/columns': columns,
   '/symbols': symbols,
+  '/js-code': JS_CODE_CONTENT,
   '/symbols-without-content': CONTENT_WITHOUT_SYMBOLS,
   '/symbol-bindings': symbolBindings,
   '/symbol-with-locale': symbolWithLocale,
@@ -61,6 +66,7 @@ const pages = {
   '/custom-breakpoints': customBreakpoints,
   '/reactive-state': reactiveState,
   '/element-events': elementEvents,
+  '/external-data': EXTERNAL_DATA,
   '/show-hide-if': SHOW_HIDE_IF,
   '/show-hide-if-repeats': SHOW_HIDE_IF_REPEATS,
   '/custom-breakpoints-reset': customBreakpointsReset,
@@ -71,6 +77,8 @@ const pages = {
   '/video': video,
   '/repeat-items-bindings': REPEAT_ITEMS_BINDINGS,
   '/input-default-value': INPUT_DEFAULT_VALUE,
+  '/duplicate-attributes': DUPLICATE_ATTRIBUTES,
+  '/js-content-is-browser': JS_CONTENT_IS_BROWSER,
 } as const;
 
 const apiVersionPathToProp = {
@@ -79,13 +87,13 @@ const apiVersionPathToProp = {
   '/api-version-v3': { apiVersion: 'v3' },
 } as const;
 
-export type Path = keyof typeof pages;
+export type Path = keyof typeof PAGES;
 
 const GEN1_ONLY_PATHNAMES: Path[] = ['/api-version-v1'];
 const GEN2_ONLY_PATHNAMES: Path[] = ['/api-version-v2'];
 
 export const getAllPathnames = (target: 'gen1' | 'gen2'): string[] => {
-  return Object.keys(pages).filter(pathname => {
+  return Object.keys(PAGES).filter(pathname => {
     if (target === 'gen1') {
       return !GEN2_ONLY_PATHNAMES.includes(pathname as Path);
     } else {
@@ -95,7 +103,7 @@ export const getAllPathnames = (target: 'gen1' | 'gen2'): string[] => {
 };
 
 const getContentForPathname = (pathname: string): BuilderContent | null => {
-  return pages[pathname as keyof typeof pages] || null;
+  return PAGES[pathname as keyof typeof PAGES] || null;
 };
 
 // remove trailing slash from pathname if it exists