+++ packages/sdks/overrides/vue/vue2/src/constants/target.ts
@@ -1,4 +1,0 @@
-type Target = import('../types/targets.js').Target;
-
-/** This file should be overriden for each framework. Ideally this would be implemented in Mitosis.  */
-export const TARGET = 'vue2' as Target;