+++ packages/sdks/src/components/blocks/blocks.lite.tsx
@@ -1,4 +1,3 @@
-import type { Signal } from '@builder.io/mitosis';
 import {
   For,
   Show,
@@ -8,22 +7,9 @@ import {
 } from '@builder.io/mitosis';
 import BuilderContext from '../../context/builder.context.lite.js';
 import ComponentsContext from '../../context/components.context.lite.js';
-import type {
-  BuilderContextInterface,
-  RegisteredComponents,
-} from '../../context/types.js';
 import Block from '../block/block.lite.jsx';
-import BlockStyles from '../block/components/block-styles.lite.jsx';
-import type { BlocksWrapperProps } from './blocks-wrapper.lite.jsx';
 import BlocksWrapper from './blocks-wrapper.lite.jsx';
-
-export type BlocksProps = Partial<
-  Omit<BlocksWrapperProps, 'BlocksWrapper' | 'BlocksWrapperProps'>
-> & {
-  context?: Signal<BuilderContextInterface>;
-  registeredComponents?: RegisteredComponents;
-  linkComponent: any;
-};
+import type { BlocksProps } from './blocks.types.js';
 
 useMetadata({
   rsc: {
@@ -44,15 +30,11 @@ export default function Blocks(props: BlocksProps) {
       BlocksWrapper={props.context?.value?.BlocksWrapper}
       BlocksWrapperProps={props.context?.value?.BlocksWrapperProps}
     >
-      {/**
-       * We need to run two separate loops for content + styles to workaround the fact that Vue 2
-       * does not support multiple root elements.
-       */}
       <Show when={props.blocks}>
         <For each={props.blocks}>
           {(block) => (
             <Block
-              key={'render-block-' + block.id}
+              key={block.id}
               block={block}
               context={useTarget({
                 rsc: props.context,
@@ -65,20 +47,6 @@ export default function Blocks(props: BlocksProps) {
                   componentsContext.registeredComponents,
               })}
               linkComponent={props.linkComponent}
-            />
-          )}
-        </For>
-      </Show>
-      <Show when={props.blocks}>
-        <For each={props.blocks}>
-          {(block) => (
-            <BlockStyles
-              key={'block-style-' + block.id}
-              block={block}
-              context={useTarget({
-                rsc: props.context?.value,
-                default: props.context?.value || builderContext.value,
-              })}
             />
           )}
         </For>