+++ packages/sdks/src/functions/get-block-properties.ts
@@ -32,8 +32,7 @@ function getStyleAttribute(
 ): string | Partial<CSSStyleDeclaration> {
   switch (TARGET) {
     case 'svelte':
-    case 'vue2':
-    case 'vue3':
+    case 'vue':
     case 'solid':
       return convertStyleMapToCSSArray(style).join(' ');
     case 'qwik':